Add clearEnrollmentState feature to Tiqr Service
This allows us to reset the enrollment state before a registration
process starts. This is a requirement when multiple registrations are
started in the users browser (in multiple tabs or browser windows). The
browser php session state would collide with the application state (PDO,
Memcache or filesystem) when multiple enrollment sessions where active.