Remove LDAP support from User(Secret)Storage
Removed support for creating an instance of a custom storage solution
by providing the FQN of the implementation as $type. This construction
makes the library more open for implementation. If ever somebody wants
to support a different storage solution, they can add it to the library,
by creating a concrete solution to the library + updating the factory.