Remove OCRAWrappers
The library contains two OCRA implementations with for each an OCRAWrappers proxy interface. For use with the Tiqr client it was verified that these two implementations can be used interchangebly. The "v2" implementation did previously contain a bug in the response code generation, which has been fixed a while ago. The only significatnt difference now is that the generateSessionData functions generate 32 vs 64 (depends on the OCRA suite, 64 for the default ) hex digit long strings.

Removing the OCRAWrappers and one of the OCRA implementations makes a lot of sence. We keep the new RFC reference based implementation.

The OcraService was changed to use the new OCRA implementation. Session key generation for OCRA and other purposes was moved to the Tiqr_Service class, and is now independed of the OCRASuite.

The oathservice OcraService was updated as well, but is not tested.